requirements.txt
# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xdist==3.5.0

# HTTP client
requests==2.31.0
httpx==0.25.2

# Docker management
docker==7.0.0
python-dotenv==1.0.0

# Performance testing
pytest-benchmark==4.0.0
locust==2.20.0

# Security testing
python-owasp-zap-v2.4==0.0.21

# Data validation
jsonschema==4.20.0
faker==20.1.0

# JWT handling
pyjwt==2.8.0
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6

# Coverage and reporting
pytest-cov==4.1.0
pytest-html==4.1.1
allure-pytest==2.13.2
